// apps/foreldrepengesoknad/src/app/types/AttachmentType.ts
export enum AttachmentType {
    OMSORGSOVERTAKELSE = 'omsorgsovertakelse',
    ADOPSJONSVEDTAK = 'adopsjonsvedtak',
    TERMINBEKREFTELSE = 'terminbekreftelse',
    FØDSELSATTEST = 'fødselsattest',
    ANNEN_INNTEKT = 'anneninntektDokumentasjon',
    UTSETTELSE_SYKDOM = 'utsettelseSykdomUttaksplan',
    MORS_AKTIVITET_DOKUMENTASJON = 'morsaktivitetdokumentasjon',
    OVERFØRING_KVOTE = 'dokumentasjonOverføringAvKvote',
    ALENEOMSORG = 'dokumentasjonAvAleneomsorg',
    SEN_ENDRING = 'senEndring',
    HV_ØVELSE = 'hvØvelse',
    NAV_TILTAK = 'navTiltak',
}